pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ecomscrape"
version = "0.1.0"
description = "Config-driven e-commerce scraping framework with an optional API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31.0",
  "beautifulsoup4>=4.12.0",
  "lxml>=4.9.0",
  "pandas>=2.2.0",
  "openpyxl>=3.1.0",
  "pyyaml>=6.0.0",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
test = ["pytest>=8.0.0"]

[project.scripts]
ecomscrape = "ecomscrape.cli:main"

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*"]
